Run build every Sunday at 4AM
The reasoning for this is that even if nothing changes in this project
the build will still run and we will be notified of failures earlier
rather than finding out just the next time we make changes.

Sometimes external factors can cause the build to fail even though we
haven't changed anything. For instance, it appears that there was an
issue with a recent build due to us using an older version of the ubuntu
base image.