Support specifying git branch
Adds support for the user to specify what git branch they want to clone
from when building.

This should likely be used along with a different tag than 'latest' so
images not built on the mainline branch aren't pushed to latest. Docker
tags should still be specified with the image name.